refactor: change the name of import
