+++ front/components/forms/login-form.tsx
@@ -4,7 +4,7 @@ import { useAuth } from "../../context/AuthContext"
 import { initStateSingIn } from "../../types/auth"
 import { useActionState, useEffect } from "react"
 import { CustomLink } from "../link/customLink"
-import { userSingIn } from "../../actions/auth"
+import { userSingIn } from "../../lib/actions/auth"
 import { CustomInput } from "./parts/input"
 import { useRouter } from "next/navigation"
 import { Button } from "./parts/button"
